refactored new_config in preparation for schema change checks
