+++ hack/new_version.sh
@@ -16,7 +16,7 @@
 
 set -e
 
-go run ./hack/new_config_version/version.go $@
+go run ./hack/versions/cmd/new/version.go $@
 
 goimports -w ./pkg/skaffold/schema
 make generate-schemas
@@ -26,7 +26,7 @@ make test
 echo
 echo "---------------------------------------"
 echo
-echo "Files generated for $NEW_VERSION."
+echo "Files generated for the new version."
 echo "All tests should have passed. For the docs change, commit the results and rerun 'make test'."
 echo "Please double check manually the generated files as well: the upgrade functionality, and all the examples:"
 echo
